+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes for Elasticsearch Plugin
 
-## Unreleased
+## 2.1.0
 
 ### Added